call DispatchMessage in event loop and combine handle_message with wnd_proc
