+++ src/win/mod.rs
@@ -1,4 +1,3 @@
-mod event;
 mod window;
 
 pub use window::*;
